% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_matrix}
\alias{check_matrix}
\title{Create Check Matrix}
\usage{
check_matrix(x)
}
\arguments{
\item{x}{\code{list} results of tests.}
}
\value{
\code{data.table} with test results as columns.
}
\description{
Create Check Matrix
}
