% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compress_detailed_tests}
\alias{compress_detailed_tests}
\title{Compress Results of Detailed Tests}
\usage{
compress_detailed_tests(dt)
}
\arguments{
\item{dt}{\code{list} results of detailed tests.}
}
\value{
\code{list} with test failures.
}
\description{
Subsets results of the tests, where at least one row failed.
}
