% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{concatenate_test_failures}
\alias{concatenate_test_failures}
\title{Concatenate Validation Test Failures Descriptions}
\usage{
concatenate_test_failures(test_failures)
}
\arguments{
\item{test_failures}{\code{data.frame} with test results as columns.}
}
\value{
\code{character} concatenated descriptions of test failures with
one string pr. row.
}
\description{
Concatenates validation test failures descriptions to a single character
vector.
}
