% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_test_results_df}
\alias{create_test_results_df}
\title{Create Data Frame with Test Results}
\usage{
create_test_results_df(x)
}
\arguments{
\item{x}{\code{list} results of tests.}
}
\value{
\code{data.table} with test results as columns.
}
\description{
Create Data Frame with Test Results
}
