% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_tests_meta_data}
\alias{create_tests_meta_data}
\title{Create Meta Data of Validation Tests}
\usage{
create_tests_meta_data()
}
\value{
\code{list} meta data of validation tests.
}
\description{
Creates meta data of available validation tests as a list. The list has as
many elements as the number of available validation test - one for each test.
Entries are named after the different tests.
}
\details{
The meta data of a validation test consists of:

\describe{
  \item{evaluate_level}{is the test evaluated on column level (`col`) or on
  row level (`row`)?}
  \item{evaluate_class}{what classes of variables are being tested with this
  specific test?}
  \item{description}{a short description of what a test failure means for
  the given test}
}
}
\examples{
create_tests_meta_data()
}
