% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_rows.R
\name{get_clean_rows}
\alias{get_clean_rows}
\title{Get Clean Rows}
\usage{
get_clean_rows(playback, ignore_check_names = NULL, ignore_cols = NULL,
  ignore_combinations = NULL)
}
\arguments{
\item{playback}{\code{data.playback} to generate violation matrix from.}

\item{ignore_check_names}{\code{character} with names of checks to ignore.}

\item{ignore_cols}{\code{character} with names of columns/variables to ignore.}

\item{ignore_combinations}{\code{list} with combinations of checks and
columns to ignore.}
}
\value{
\code{logical} with the same length as the number of rows in new 
data. The value is TRUE, if the row passed all checks, otherwise FALSE.
}
\description{
Get Clean Rows
}
