% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_test_descriptions}
\alias{get_test_descriptions}
\title{Get Descriptions of Validation Tests}
\usage{
get_test_descriptions()
}
\value{
\code{data.frame} test descriptions.
}
\description{
Get Descriptions of Validation Tests
}
\examples{
get_test_descriptions()
}
