% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_tests_meta_data}
\alias{get_tests_meta_data}
\title{Get Meta Data of Validation Tests in a Data Frame}
\usage{
get_tests_meta_data()
}
\value{
\code{data.frame} meta data of validation tests.
}
\description{
Gets meta data of available validation tests as a data.frame.
}
\details{
The meta data of a validation test consists of:

\describe{
  \item{test_name}{name of the test}
  \item{evaluate_level}{is the test evaluated on column level (`col`) or on
  row level (`row`)?}
  \item{evaluate_class}{what classes of variables are being tested with this
  specific test?}
  \item{description}{a short description of what a test failure means for
  the given test}
}
}
\examples{
get_tests_meta_data()
}
