% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_rows.R
\name{get_violations}
\alias{get_violations}
\title{Get Violations}
\usage{
get_violations(playback, ignore_check_names = c("new_variable"),
  ignore_cols = NULL, ignore_combinations = list(mismatch_class =
  NULL))
}
\arguments{
\item{playback}{\code{playback} to generate violation matrix from.}

\item{ignore_check_names}{\code{character} with names of checks to ignore.}

\item{ignore_cols}{\code{character} with names of columns/variables to ignore.}

\item{ignore_combinations}{\code{list} with combinations of checks and
columns to ignore.}
}
\value{
\code{data.table} with logicals for all of the checks, that one or more
rows failed to pass. A failed check for any given row is equivalent to a value
of TRUE. If all checks passed, the function will simply return a matrix with
one column, 'any_violations', that is always FALSE, to ensure that the output
is stable and consistent.
}
\description{
Get Violations
}
