% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore}
\alias{ignore}
\title{Ignore and remove certain tests}
\usage{
ignore(tests, variables_newdata, ignore_tests = NULL,
  ignore_cols = NULL, ignore_combinations = NULL)
}
\arguments{
\item{tests}{\code{list} with computed tests.}

\item{variables_newdata}{\code{character} names of variables in new data.}

\item{ignore_tests}{\code{character} with names of tests to ignore.}

\item{ignore_cols}{\code{character} with names of columns/variables to ignore.}

\item{ignore_combinations}{\code{list} with combinations of tests and
columns to ignore.}
}
\value{
\code{list} with only the relevant tests.
}
\description{
Ignore and remove certain tests
}
