% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore_cols}
\alias{ignore_cols}
\title{Ignore Test Results from Tests of Specific Columns}
\usage{
ignore_cols(tests, col_names, variables_newdata)
}
\arguments{
\item{tests}{\code{list} test results.}

\item{col_names}{\code{character} names of columns for which test results
should be ignored.}

\item{variables_newdata}{\code{character} names of variables in new data.}
}
\value{
\code{list} results after removing tests.
}
\description{
Ignore Test Results from Tests of Specific Columns
}
