% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore_combinations}
\alias{ignore_combinations}
\title{Ignore Test Results from Specific Tests of Specific Columns}
\usage{
ignore_combinations(tests, combinations, variables_newdata)
}
\arguments{
\item{tests}{\code{list} test results.}

\item{combinations}{\code{list} combinations of tests and columns from which
test results should be ignored.}

\item{variables_newdata}{\code{character} names of variables in new data.}
}
\value{
\code{list} test results after removals.
}
\description{
Ignore Test Results from Specific Tests of Specific Columns
}
