% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore_tests}
\alias{ignore_tests}
\title{Ignore Results from Specific Tests}
\usage{
ignore_tests(tests, test_names = NULL)
}
\arguments{
\item{tests}{\code{list} test results.}

\item{test_names}{\code{character} names of tests to be ignored.}
}
\value{
\code{list} results after removing specific tests.
}
\description{
Ignore Results from Specific Tests
}
