% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_newdata}
\alias{iris_newdata}
\title{Simulated Iris New Data}
\format{A data.frame with 150 rows and 5 columns.}
\source{
Script attached.
}
\usage{
iris_newdata
}
\description{
A mutated version of the famous `iris` data set.
}
\keyword{datasets}
