% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{order_by_tests}
\alias{order_by_tests}
\title{Order Test Results by Test Names}
\usage{
order_by_tests(dt)
}
\arguments{
\item{dt}{\code{list} test results.}
}
\value{
\code{list} test results ordered by test names.
}
\description{
Order Test Results by Test Names
}
