% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play.data.frame}
\alias{play.data.frame}
\title{Validate New Data by Playing a Data Tape on Them}
\usage{
\method{play}{data.frame}(x, tape, verbose = TRUE, ...)
}
\arguments{
\item{x}{\code{data.frame} new data.}

\item{tape}{\code{data.tape} parameters and meta data recorded from 
training data.}

\item{verbose}{\code{logical} should messages be printed?}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{data.playback} results from validations tests. The list will 
inherit from the \code{data.playback}.
}
\description{
Runs a set of validation tests on new data for an existing machine learning 
model. These tests are based on parameters recorded (with 
\code{\link{record}}) from the training data.
}
