% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play.default}
\alias{play.default}
\title{Validate Variable}
\usage{
\method{play}{default}(x, tape, ...)
}
\arguments{
\item{x}{anything.}

\item{tape}{\code{list} parameters and meta data recorded from training data.}

\item{...}{all further arguments.}
}
\value{
\code{list} results from validations tests.
}
\description{
Runs a set of validation tests on variable in new data. These tests
are based on parameters recorded (with \code{\link{record}}) from the 
training data.
}
