% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play.integer}
\alias{play.integer}
\title{Validate Integer}
\usage{
\method{play}{integer}(x, tape, ...)
}
\arguments{
\item{x}{\code{integer} from new data.}

\item{tape}{\code{list} parameters and meta data recorded from training data.}

\item{...}{all further arguments.}
}
\value{
\code{list} results from validations tests.
}
\description{
Runs a set of validation tests on a \code{integer} in new data. These tests
are based on parameters recorded (with \code{\link{record}}) from the 
training data.
}
