% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play Data Recording on a New Data Set}
\usage{
play(x, ...)
}
\arguments{
\item{x}{\code{data.frame} data set to validate.}

\item{...}{all other arguments.}
}
\value{
\code{list} relevant statistics used for checking a new data set.
}
\description{
Play Data Recording on a New Data Set
}
