% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.data.playback}
\alias{print.data.playback}
\title{Print Data Playback}
\usage{
\method{print}{data.playback}(x, ...)
}
\arguments{
\item{x}{A `data.playback` object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The original object (invisibly)
}
\description{
Print Data Playback
}
\examples{
# record tape from `iris`.
tape <- record(iris)
# load data.
data(iris_newdata)
# validate new data by playing new tape on it.
playback <- play(tape, iris_newdata)
# print it.
print(playback)
}
