% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.playback}
\alias{print.playback}
\title{Print a Playback}
\usage{
\method{print}{playback}(x, first = 10)
}
\arguments{
\item{x}{A `playback` object}

\item{first}{The number of row indices to print pr. check.}
}
\value{
The original object (invisibly)
}
\description{
Print a Playback
}
\author{
Lars Kjeldgaard
}
