% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.data.frame}
\alias{record.data.frame}
\title{Record Parameters and Meta Data from Data Frame}
\usage{
\method{record}{data.frame}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{\code{data.frame} training data for machine learning model.}

\item{verbose}{\code{logical} should messages be printed?}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded parameters and meta data.
}
\description{
Records parameters and meta data from a data.frame.
}
\examples{
record(iris)
}
