% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.factor}
\alias{record.factor}
\title{Record Parameters and Meta Data from Factor}
\usage{
\method{record}{factor}(x, ...)
}
\arguments{
\item{x}{\code{factor}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded parameters and meta data.
}
\description{
Records parameters and meta data from \code{factor}.
}
\examples{
record(iris$Species)
}
