% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.integer}
\alias{record.integer}
\title{Record Parameters and Meta Data from Integer}
\usage{
\method{record}{integer}(x, ...)
}
\arguments{
\item{x}{\code{integer}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded parameters and meta data.
}
\description{
Records parameters and meta data from vector with integers.
}
\examples{
record(c(1:10, NA_integer_))
}
