% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.numeric}
\alias{record.numeric}
\title{Record Parameters and Meta Data from Numeric}
\usage{
\method{record}{numeric}(x, ...)
}
\arguments{
\item{x}{\code{numeric}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded parameters and meta data.
}
\description{
Records parameters and meta data from numeric vector.
}
\examples{
record(iris$Sepal.Length)
}
