% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record Parameters and Meta Data from Training Data}
\usage{
record(x, ...)
}
\arguments{
\item{x}{training data or just a single variable to record parameters and 
other relevant metadata from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{list} recorded parameters and metadata. The list will inherit
from the \code{data.tape} class when the function is invoked with a 
\code{data.frame}.
}
\description{
Records relevant meta data and parameters from training data for a machine 
learning model. The recorded data can then be used to compute a set of 
validation tests on new data with \code{\link{play}}.
}
\examples{
record(iris)
}
