% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{run_validation_tests.numeric}
\alias{run_validation_tests.numeric}
\title{Run Validation Tests on Numeric}
\usage{
\method{run_validation_tests}{numeric}(x, parameters, ...)
}
\arguments{
\item{x}{\code{numeric} in new data.}

\item{parameters}{\code{list} parameters and meta data of the same variable
recorded from training data (with \code{\link{record}}).}

\item{...}{further arguments passed to or from other methods. Not used at
the moment.}
}
\value{
\code{list} results from validations tests.
}
\description{
Runs a set of validation tests on a \code{numeric} in new data. These tests
are based on parameters recorded (with \code{\link{record}}) from the 
training data.
}
