% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_newdata_iris.R
\name{simulate_newdata_iris}
\alias{simulate_newdata_iris}
\title{Simulate New Data from Iris Data Set}
\usage{
simulate_newdata_iris()
}
\value{
\code{data.frame} simulated data set.
}
\description{
Simulate New Data from Iris Data Set
}
\examples{
simulate_newdata_iris()
}
