% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tests_meta_data}
\alias{tests_meta_data}
\title{Meta Data of Validation Tests}
\format{A list with 7 entries, one for each validation test. Entries are 
named after the different tests. For each validation test there are two
pieces of information available:
\describe{
  \item{evaluate_level}{is the test evaluated on column level (`col`) or on
  row level (`row`)?}
  \item{description}{a short description of what a test failure means for
  the given test}
}}
\source{
Script attached.
}
\usage{
tests_meta_data
}
\description{
A dataset containing relevant information of the validation tests.
}
\keyword{datasets}
