% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_rows.R
\name{get_violations_string}
\alias{get_violations_string}
\title{Get Violations as a String}
\usage{
get_violations_string(playback, ignore_test_names = c("new_variable"),
  ignore_cols = NULL, ignore_combinations = list(mismatch_class =
  NULL))
}
\value{
\code{character} with one entry for each row in new data. Each
entry concatenates information of the tests, that did NOT pass for the
corresponding row in new data.
}
\description{
Get Violations as a String
}
\details{
- 'mismatch_levels': (only relevant for factors and characters).
- 'missing_variable': the variable was recorded on training data but not 
observed in new data.
}
