% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore}
\alias{ignore}
\title{Ignore and remove certain checks}
\usage{
ignore(checks, variables_newdata, ignore_check_names = NULL,
  ignore_cols = NULL, ignore_combinations = NULL)
}
\arguments{
\item{checks}{\code{list} with computed checks.}

\item{variables_newdata}{\code{character} names of variables in new data.}

\item{ignore_check_names}{\code{character} with names of checks to ignore.}

\item{ignore_cols}{\code{character} with names of columns/variables to ignore.}

\item{ignore_combinations}{\code{list} with combinations of checks and
columns to ignore.}
}
\value{
\code{list} with only the relevant checks.
}
\description{
Ignore and remove certain checks
}
