% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore}
\alias{ignore}
\title{Ignore Certain Test Results}
\usage{
ignore(tests, variables_newdata, ignore_tests = NULL,
  ignore_cols = NULL, ignore_combinations = NULL)
}
\arguments{
\item{tests}{\code{list} test results.}

\item{variables_newdata}{\code{character} names of variables in new data.}

\item{ignore_tests}{\code{character} ignore test results from tests with 
these names.}

\item{ignore_cols}{\code{character} ignore test results from tests of 
columns with these names.}

\item{ignore_combinations}{\code{list} ignore test results from specific 
tests of specific columns.}
}
\value{
\code{list} only the relevant test results.
}
\description{
Ignores certain test results in accordance with user inputs.
}
\details{
Look up the descriptions and other meta data of the available
validation tests with \code{\link{get_tests_meta_data}}.
}
