% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Validate New Data by Playing a Data Tape on It}
\usage{
play(tape, newdata, verbose = TRUE)
}
\arguments{
\item{tape}{\code{data.tape} parameters and meta data recorded from 
training data.}

\item{newdata}{\code{data.frame} new data to be predicted with an existing 
machine learning model.}

\item{verbose}{\code{logical} should messages be printed?}
}
\value{
\code{data.playback} results from validations tests.
}
\description{
Runs a set of validation tests on new data to be predicted with an existing 
machine learning model. These tests are based on parameters recorded (with 
\code{\link{record}}) from training data.
}
\examples{
# record tape from `iris`.
tape <- record(iris)
# simulate new data.
newdata <- simulate_newdata_iris()
# validate new data by playing new tape on it.
play(tape, newdata)
}
