% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.playback}
\alias{print.playback}
\title{Print Playback}
\usage{
\method{print}{playback}(x, ...)
}
\arguments{
\item{x}{A `playback` object.}

\item{first}{The number of row indices to print pr. failed check.}
}
\value{
The original object (invisibly)
}
\description{
Print Playback
}
\author{
Lars Kjeldgaard
}
