% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.playback}
\alias{print.playback}
\title{Print a Playback}
\usage{
\method{print}{playback}(x, form_width = 30, ...)
}
\arguments{
\item{x}{A `playback` object}

\item{form_width}{The number of characters used to print the variables or
terms in a formula}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\value{
The original object (invisibly)
}
\description{
Print a Playback
}
\author{
Lars Kjeldgaard
}
