% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.character}
\alias{record.character}
\title{Record Parameters and Meta Data from Character}
\usage{
\method{record}{character}(x, ...)
}
\arguments{
\item{x}{\code{character}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded parameters and meta data. The unique values
of the vector are recorded as `levels`.
}
\description{
Records parameters and meta data from character.
}
\examples{
record(letters)
}
