% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.character}
\alias{record.character}
\title{Record Statistics and Meta Data of a Character}
\usage{
\method{record}{character}(x, ...)
}
\arguments{
\item{x}{\code{character}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded statistics and meta data.
}
\description{
Records statistics and meta data of a \code{character}.
}
\examples{
record(letters)
}
