% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.data.frame}
\alias{record.data.frame}
\title{Record Statistics and Meta Data of a Data Frame}
\usage{
\method{record}{data.frame}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{\code{data.frame} training data for predictive model.}

\item{verbose}{\code{logical} should messages be printed?}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded statistics and meta data.
}
\description{
Records Statistics and meta data of a data.frame.
}
\examples{
record(iris)
}
