% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.factor}
\alias{record.factor}
\title{Record Statistics and Meta Data of a Factor}
\usage{
\method{record}{factor}(x, ...)
}
\arguments{
\item{x}{\code{factor}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded statistics and meta data.
}
\description{
Records statistics and meta data of a \code{factor}.
}
\examples{
record(iris$Species)
}
