% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.integer}
\alias{record.integer}
\title{Record Statistics and Meta Data of an Integer}
\usage{
\method{record}{integer}(x, ...)
}
\arguments{
\item{x}{\code{integer}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded statistics and meta data.
}
\description{
Records statistics and meta data of an \code{integer}.
}
\examples{
record(c(1:10, NA_integer_))
}
