% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record.numeric}
\alias{record.numeric}
\title{Record Statistics and Meta Data of a Numeric}
\usage{
\method{record}{numeric}(x, ...)
}
\arguments{
\item{x}{\code{numeric}}

\item{...}{all further arguments.}
}
\value{
\code{list} recorded statistics and meta data.
}
\description{
Records statistics and meta data of a \code{numeric}.
}
\examples{
record(iris$Sepal.Length)
}
