% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record Data Statistics}
\usage{
record(x, ...)
}
\arguments{
\item{x}{what variable or dataset to record/learn statistics from.}

\item{...}{all other arguments.}
}
\value{
\code{list} relevant statistics used for validating a new data set.
}
\description{
Record Data Statistics
}
