% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record Statistics and Meta Data of Variables in Training Data}
\usage{
record(x, ...)
}
\arguments{
\item{x}{training data (or just a single variable from the training data) to 
record the statistics and other relevant meta data of.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{list} recorded statistics and meta data. The list will inherit
from the \code{data.tape} class when the function is invoked with a 
\code{data.frame}.
}
\description{
Records statistics and meta data of variables in the training data for a 
predictive model. The recorded data can then be used to compute a set 
of validation tests on new data with \code{\link{play}}.
}
\examples{
record(iris)
}
